anykernel.sh: Support both SDCARDFS & NON-SDCARDFS in the same zip [2]
We can build both images and include them in the zip then dynamically
move them to where they need to go.

Previously we were checking with /proc/config.gz which turned out to be
unreliable at times due to the fact that not all sdcardfs kernels point
to a sdcardfs-enabled stock defconfig for config.gz.

Thanks to @YaroST12 for suggesting to use mount to find if its a
sdcardfs rom or not instead of the /proc/config.gz method.

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>